# docs/contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Digital Procurement and PLS Seedbed API
  version: 0.24.0
  summary: Compliance-first procurement evidence API with proof anchoring and PLS seedbed support.
  description: >
    API contract for the supervisor-demo plus selected pilot-hardening baseline.
    The runtime uses opaque bearer sessions. OAuth2/OIDC entries are readiness
    boundaries and do not claim a production identity provider.
...
servers:
  - url: http://localhost:3100/api/v1
    description: Local backend
...
paths:
...
  /source-to-award/requisitions:
    post:
      tags: [Source to Award]
      operationId: createSourceToAwardRequisition
      summary: Create a sourcing requisition for buyer approval.
      security:
        - bearerSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, estimatedAmount, currency]
              properties:
                title: { type: string }
                description: { type: string }
                estimatedAmount: { type: string, example: "68000.00" }
                currency: { type: string, example: MYR }
      responses:
        "200":
          description: Requisition created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceToAwardResponse"
  /source-to-award/requisitions/{id}/approve:
    post:
      tags: [Source to Award]
      operationId: approveSourceToAwardRequisition
      summary: Approve a requisition for RFQ issuance.
      security:
        - bearerSession: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Requisition approved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceToAwardResponse"
  /source-to-award/rfqs:
    post:
      tags: [Source to Award]
      operationId: issueSourceToAwardRfq
      summary: Issue an RFQ to invited supplier organizations.
      security:
        - bearerSession: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [requisitionId, supplierOrganizationIds]
              properties:
                requisitionId: { type: string }
                supplierOrganizationIds:
                  type: array
                  items: { type: string }
                responseDeadline: { type: string, format: date-time }
      responses:
        "200":
          description: RFQ issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceToAwardResponse"
  /source-to-award/rfqs/{id}/quotations:
    post:
      tags: [Source to Award]
      operationId: submitSourceToAwardQuotation
      summary: Submit a supplier quotation for an invited RFQ.
      security:
        - bearerSession: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount, currency]
              properties:
                amount: { type: string, example: "68000.00" }
                currency: { type: string, example: MYR }
                deliveryDays: { type: integer }
                notes: { type: string }
      responses:
        "200":
          description: Quotation submitted.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceToAwardResponse"
  /source-to-award/rfqs/{id}/award:
    post:
      tags: [Source to Award]
      operationId: awardSourceToAwardRfq
      summary: Select a quotation and generate the purchase order handoff.
      security:
        - bearerSession: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [quotationId]
              properties:
                quotationId: { type: string }
                rationale: { type: string }
      responses:
        "201":
          description: Award selected and order generated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceToAwardResponse"
  /source-to-award/cases:
    get:
      tags: [Source to Award]
      operationId: listSourceToAwardCases
      summary: List source-to-award cases visible to the authenticated actor.
      security:
        - bearerSession: []
      responses:
        "200":
          description: Visible source-to-award cases.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceToAwardListResponse"
  /source-to-award/cases/{caseId}:
    get:
      tags: [Source to Award]
      operationId: getSourceToAwardCase
      summary: Get a source-to-award case.
      security:
        - bearerSession: []
      parameters:
        - name: caseId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Source-to-award case.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceToAwardResponse"
...
components:
  securitySchemes:
    bearerSession:
      type: http
      scheme: bearer
      bearerFormat: opaque-session
      description: Current runtime opaque bearer session token.
    oauth2:
      type: oauth2
      description: OAuth2/OIDC readiness boundary only; no production IdP is configured.
      flows:
        authorizationCode:
          authorizationUrl: https://identity.example.test/oauth/authorize
          tokenUrl: https://identity.example.test/oauth/token
          scopes:
            org:read: Read organization profile and context.
            org:write: Update organization metadata.
            org:users:manage: Manage organization users.
            network:read: Read organization network graph.
            network:write: Create and decide network requests.
            deals:read: Read company ledger projections.
            deals:write: Prepare deal actions.
            productivity:read: Read productivity workspace.
            productivity:write: Update tasks and saved views.
            proof:read: Read proof metadata.
            proof:verify: Verify proof hashes.
            email:read: Read local notification outbox.
            admin:platform: Platform administration.
...
  schemas:
...
    SourceToAwardCase:
      type: object
      properties:
        caseId: { type: string }
        status: { type: string }
        requisition: { type: object }
        rfq: { type: object }
        quotations:
          type: array
          items: { type: object }
        award: { type: object }
        generatedOrderId: { type: string }
        lifecycleEventIds:
          type: array
          items: { type: string }
        latestLifecyclePayloadHash: { type: string, example: sha256:abc123 }
    SourceToAwardResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [case]
          properties:
            case:
              $ref: "#/components/schemas/SourceToAwardCase"
    SourceToAwardListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/SourceToAwardCase"
...